requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
ruamel.yaml==0.15.37
pandas-profiling
jupyter
sphinx_rtd_theme
pandas
numpy
matplotlib
seaborn
altair
scikit-learn
selenium